+++ admb/branches/trunk-issue-79/src/linad99/multiindex.cpp
@@ -62,7 +62,7 @@
 #  if (__MSVC32__>=8) || defined(__SUNPRO_CC)
       offset+=int(pow(double(sz),i-imin))*(index(i)-mind);
 #  else
-      offset+=pow(sz,i-imin)*(index(i)-mind);
+      offset+=int(pow(sz,i-imin))*(index(i)-mind);
 #  endif
     }
     return offset;
